+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+* highlight newly added question ([PR #28](https://github.com/liveask/liveask/pull/28))
+
 ## [2.3.0/2.3.1] - 2022-07-25
 
 ### Added